refactor jitted taumol*into radsw_bands.py
